+++ include/pigo/impl/pigo.impl.hpp
@@ -137,6 +137,8 @@ namespace pigo {
             return PIGO_COO_BIN;
         if (r.at_str(CSR<>::csr_file_header))
             return PIGO_CSR_BIN;
+        if (r.at_str(DiGraph<>::digraph_file_header))
+            return PIGO_DIGRAPH_BIN;
         if (r.at_str("PIGO"))
             throw Error("Unsupported PIGO binary format, likely version mismatch");
         // Check the filename for .mtx