+++ tests/coo/test_coo_sym.cpp
@@ -45,11 +45,11 @@ int read_clean(string dir_path) {
 
 int read_with_sl(string dir_path) {
     COO<int, int, vector<int>, true> c { dir_path + "/selfloop.el" };
-    EQ(c.m(), 20);
+    EQ(c.m(), 17);
     EQ(c.n(), 4);
 
-    vector<int> valid_x = {0,1, 0,2, 0,3, 1,0, 1,1, 2,1, 2,2, 3,0, 3,2, 3,3};
-    vector<int> valid_y = {1,0, 2,0, 3,0, 0,1, 1,1, 1,2, 2,2, 0,3, 2,3, 3,3};
+    vector<int> valid_x = {0,1, 0,2, 0,3, 1,0, 1, 2,1, 2, 3,0, 3,2, 3};
+    vector<int> valid_y = {1,0, 2,0, 3,0, 0,1, 1, 1,2, 2, 0,3, 2,3, 3};
 
     NOPRINT_EQ(valid_x, c.x());
     NOPRINT_EQ(valid_y, c.y());
@@ -73,6 +73,64 @@ int read_without_sl(string dir_path) {
     return 0;
 }
 
+int weight_sym(string dir_path) {
+    WCOOPtr<int, int, double, true> c {dir_path+"/weighted.mtx"};
+
+    // Ensure that the count is correct
+    EQ(c.m(), 12);
+    EQ(c.n(), 6);
+
+    auto x = c.x();
+    auto y = c.y();
+    auto w = c.w();
+
+    size_t ctr = 0;
+    EQ(x[ctr], 5); EQ(y[ctr], 1); FEQ(w[ctr++], 4.3333);
+    EQ(y[ctr], 5); EQ(x[ctr], 1); FEQ(w[ctr++], 4.3333);
+    EQ(x[ctr], 4); EQ(y[ctr], 1); FEQ(w[ctr++], 0.00003454);
+    EQ(y[ctr], 4); EQ(x[ctr], 1); FEQ(w[ctr++], 0.00003454);
+    EQ(x[ctr], 4); EQ(y[ctr], 2); FEQ(w[ctr++], 3);
+    EQ(y[ctr], 4); EQ(x[ctr], 2); FEQ(w[ctr++], 3);
+    EQ(x[ctr], 4); EQ(y[ctr], 1); FEQ(w[ctr++], 1e10);
+    EQ(y[ctr], 4); EQ(x[ctr], 1); FEQ(w[ctr++], 1e10);
+    EQ(x[ctr], 2); EQ(y[ctr], 1); FEQ(w[ctr++], -0.04321);
+    EQ(y[ctr], 2); EQ(x[ctr], 1); FEQ(w[ctr++], -0.04321);
+    EQ(x[ctr], 3); EQ(y[ctr], 1); FEQ(w[ctr++], -6.6E-10);
+    EQ(y[ctr], 3); EQ(x[ctr], 1); FEQ(w[ctr++], -6.6E-10);
+
+    c.free();
+
+    return 0;
+}
+
+int intweight_sym(string dir_path) {
+    WCOO<int, int, shared_ptr<int>, int, shared_ptr<int>, true>
+        c { dir_path + "/intweight.mtx" };
+
+    EQ(c.m(), 10);
+    EQ(c.n(), 6);
+
+    auto x = c.x().get();
+    auto y = c.y().get();
+    auto w = c.w().get();
+
+    size_t ctr = 0;
+    EQ(x[ctr], 5); EQ(y[ctr], 1); EQ(w[ctr++], 4);
+    EQ(y[ctr], 5); EQ(x[ctr], 1); EQ(w[ctr++], 4);
+    EQ(x[ctr], 4); EQ(y[ctr], 1); EQ(w[ctr++], 1);
+    EQ(y[ctr], 4); EQ(x[ctr], 1); EQ(w[ctr++], 1);
+    EQ(x[ctr], 4); EQ(y[ctr], 2); EQ(w[ctr++], 3);
+    EQ(y[ctr], 4); EQ(x[ctr], 2); EQ(w[ctr++], 3);
+    EQ(x[ctr], 4); EQ(y[ctr], 1); EQ(w[ctr++], 12);
+    EQ(y[ctr], 4); EQ(x[ctr], 1); EQ(w[ctr++], 12);
+    EQ(x[ctr], 2); EQ(y[ctr], 1); EQ(w[ctr++], -10);
+    EQ(y[ctr], 2); EQ(x[ctr], 1); EQ(w[ctr++], -10);
+
+    c.free();
+
+    return 0;
+}
+
 
 int main(int argc, char **argv) {
     int pass = 0;
@@ -87,7 +145,8 @@ int main(int argc, char **argv) {
     TEST(read_clean, dir_path);
     TEST(read_with_sl, dir_path);
     TEST(read_without_sl, dir_path);
+    TEST(weight_sym, dir_path);
+    TEST(intweight_sym, dir_path);
 
     return pass;
 }
-